requirements.txt
# Python 2.7, 3.4+
attrs ~= 18.1
